+++ src/IQFeed.CSharpApiClient.Tests/Lookup/Historical/Messages/TickMessageTests.cs
@@ -20,7 +20,7 @@ namespace IQFeed.CSharpApiClient.Tests.Lookup.Historical.Messages
             var tickMessageParsed = TickMessage.Parse(message);
 
             var timestamp = DateTime.ParseExact("2018-04-17 17:51:22.123456", TickMessage.TickDateTimeFormat, CultureInfo.InvariantCulture);
-            var tickMessage = new TickMessage<decimal>(timestamp, 96.07m, 1061909, 0, 0.0m, 0.0m, 4145784264, 'O', 19, "143A");
+            var tickMessage = new TickMessage<double>(timestamp, 96.07d, 1061909, 0, 0.0d, 0.0d, 4145784264, 'O', 19, "143A");
 
             // Assert
             Assert.AreEqual(tickMessageParsed, tickMessage);
