+++ src/IQFeed.CSharpApiClient.Tests/Streaming/Level1/Messages/RegionalUpdateMessageTests.cs
@@ -17,7 +17,7 @@ namespace IQFeed.CSharpApiClient.Tests.Streaming.Level1.Messages
 
             // Act
             var regionalUpdateMessageParsed = RegionalUpdateMessage.Parse(message);
-            var regionalUpdateMessage = new RegionalUpdateMessage<decimal>("AAPL", "X", 1.23m, 100, new DateTime(2000, 1, 2, 13, 00, 00), 1.24m, 200, new DateTime(2000, 1, 2, 13, 00, 01), 0, 1, 2);
+            var regionalUpdateMessage = new RegionalUpdateMessage<double>("AAPL", "X", 1.23d, 100, new DateTime(2000, 1, 2, 13, 00, 00), 1.24d, 200, new DateTime(2000, 1, 2, 13, 00, 01), 0, 1, 2);
 
             // Arrange
             Assert.AreEqual(regionalUpdateMessageParsed, regionalUpdateMessage);
