+++ src/IQFeed.CSharpApiClient/Streaming/Level1/DynamicFieldset.cs
@@ -4,13 +4,13 @@ namespace IQFeed.CSharpApiClient.Streaming.Level1
 {
     public enum DynamicFieldset
     {
-        [FieldsetDescription("7 Day Yield", typeof(float))]
+        [FieldsetDescription("7 Day Yield", typeof(double))]
         SevenDayYield,
 
-        [FieldsetDescription("Ask", typeof(float))]
+        [FieldsetDescription("Ask", typeof(double))]
         Ask,
 
-        [FieldsetDescription("Ask Change", typeof(float))]
+        [FieldsetDescription("Ask Change", typeof(double))]
         AskChange,
 
         [FieldsetDescription("Ask Market Center", typeof(int))]
@@ -25,13 +25,13 @@ namespace IQFeed.CSharpApiClient.Streaming.Level1
         [FieldsetDescription("Available Regions", typeof(string))]
         AvailableRegions,
 
-        [FieldsetDescription("Average Maturity", typeof(float))]
+        [FieldsetDescription("Average Maturity", typeof(double))]
         AverageMaturity,
 
-        [FieldsetDescription("Bid", typeof(float))]
+        [FieldsetDescription("Bid", typeof(double))]
         Bid,
 
-        [FieldsetDescription("Bid Change", typeof(float))]
+        [FieldsetDescription("Bid Change", typeof(double))]
         BidChange,
 
         [FieldsetDescription("Bid Market Center", typeof(int))]
@@ -43,19 +43,19 @@ namespace IQFeed.CSharpApiClient.Streaming.Level1
         [FieldsetDescription("Bid Time", typeof(DateTime))]
         BidTime,
 
-        [FieldsetDescription("Change", typeof(float))]
+        [FieldsetDescription("Change", typeof(double))]
         Change,
 
-        [FieldsetDescription("Change From Open", typeof(float))]
+        [FieldsetDescription("Change From Open", typeof(double))]
         ChangeFromOpen,
 
-        [FieldsetDescription("Close", typeof(float))]
+        [FieldsetDescription("Close", typeof(double))]
         Close,
 
-        [FieldsetDescription("Close Range 1", typeof(float))]
+        [FieldsetDescription("Close Range 1", typeof(double))]
         CloseRange1,
 
-        [FieldsetDescription("Close Range 2", typeof(float))]
+        [FieldsetDescription("Close Range 2", typeof(double))]
         CloseRange2,
 
         [FieldsetDescription("Days to Expiration", typeof(string))]
@@ -70,7 +70,7 @@ namespace IQFeed.CSharpApiClient.Streaming.Level1
         [FieldsetDescription("Exchange ID", typeof(string))]
         ExchangeID,
 
-        [FieldsetDescription("Extended Trade", typeof(float))]
+        [FieldsetDescription("Extended Trade", typeof(double))]
         ExtendedTrade,
 
         [FieldsetDescription("Extended Trade Date", typeof(DateTime))]
@@ -85,10 +85,10 @@ namespace IQFeed.CSharpApiClient.Streaming.Level1
         [FieldsetDescription("Extended Trade Time", typeof(DateTime))]
         ExtendedTradeTime,
 
-        [FieldsetDescription("Extended Trading Change", typeof(float))]
+        [FieldsetDescription("Extended Trading Change", typeof(double))]
         ExtendedTradingChange,
 
-        [FieldsetDescription("Extended Trading Difference", typeof(float))]
+        [FieldsetDescription("Extended Trading Difference", typeof(double))]
         ExtendedTradingDifference,
 
         [FieldsetDescription("Financial Status Indicator", typeof(char))]
@@ -97,10 +97,10 @@ namespace IQFeed.CSharpApiClient.Streaming.Level1
         [FieldsetDescription("Fraction Display Code", typeof(string))]
         FractionDisplayCode,
 
-        [FieldsetDescription("High", typeof(float))]
+        [FieldsetDescription("High", typeof(double))]
         High,
 
-        [FieldsetDescription("Last", typeof(float))]
+        [FieldsetDescription("Last", typeof(double))]
         Last,
 
         [FieldsetDescription("Last Date", typeof(DateTime))]
@@ -118,10 +118,10 @@ namespace IQFeed.CSharpApiClient.Streaming.Level1
         [FieldsetDescription("Last Trade Date", typeof(DateTime))]
         LastTradeDate,
 
-        [FieldsetDescription("Low", typeof(float))]
+        [FieldsetDescription("Low", typeof(double))]
         Low,
 
-        [FieldsetDescription("Market Capitalization", typeof(float))]
+        [FieldsetDescription("Market Capitalization", typeof(double))]
         MarketCapitalization,
 
         [FieldsetDescription("Market Open", typeof(int))]
@@ -130,7 +130,7 @@ namespace IQFeed.CSharpApiClient.Streaming.Level1
         [FieldsetDescription("Message Contents", typeof(string))]
         MessageContents,
 
-        [FieldsetDescription("Most Recent Trade", typeof(float))]
+        [FieldsetDescription("Most Recent Trade", typeof(double))]
         MostRecentTrade,
 
         [FieldsetDescription("Most Recent Trade Conditions", typeof(string))]
@@ -148,49 +148,49 @@ namespace IQFeed.CSharpApiClient.Streaming.Level1
         [FieldsetDescription("Most Recent Trade Time", typeof(DateTime))]
         MostRecentTradeTime,
 
-        [FieldsetDescription("Net Asset Value", typeof(float))]
+        [FieldsetDescription("Net Asset Value", typeof(double))]
         NetAssetValue,
 
         [FieldsetDescription("Number of Trades Today", typeof(int))]
         NumberOfTradesToday,
 
-        [FieldsetDescription("Open", typeof(float))]
+        [FieldsetDescription("Open", typeof(double))]
         Open,
 
         [FieldsetDescription("Open Interest", typeof(int))]
         OpenInterest,
 
-        [FieldsetDescription("Open Range 1", typeof(float))]
+        [FieldsetDescription("Open Range 1", typeof(double))]
         OpenRange1,
 
-        [FieldsetDescription("Open Range 2", typeof(float))]
+        [FieldsetDescription("Open Range 2", typeof(double))]
         OpenRange2,
 
-        [FieldsetDescription("Percent Change", typeof(float))]
+        [FieldsetDescription("Percent Change", typeof(double))]
         PercentChange,
 
-        [FieldsetDescription("Percent Off Average Volume", typeof(float))]
+        [FieldsetDescription("Percent Off Average Volume", typeof(double))]
         PercentOffAverageVolume,
 
         [FieldsetDescription("Previous Day Volume", typeof(int))]
         PreviousDayVolume,
 
-        [FieldsetDescription("Price-Earnings Ratio", typeof(float))]
+        [FieldsetDescription("Price-Earnings Ratio", typeof(double))]
         PriceEarningsRatio,
 
-        [FieldsetDescription("Range", typeof(float))]
+        [FieldsetDescription("Range", typeof(double))]
         Range,
 
         [FieldsetDescription("Restricted Code", typeof(string))]
         RestrictedCode,
 
-        [FieldsetDescription("Settle", typeof(float))]
+        [FieldsetDescription("Settle", typeof(double))]
         Settle,
 
         [FieldsetDescription("Settlement Date", typeof(DateTime))]
         SettlementDate,
 
-        [FieldsetDescription("Spread", typeof(float))]
+        [FieldsetDescription("Spread", typeof(double))]
         Spread,
 
         [FieldsetDescription("Symbol", typeof(string))]
@@ -208,10 +208,10 @@ namespace IQFeed.CSharpApiClient.Streaming.Level1
         [FieldsetDescription("Type", typeof(string))]
         Type,
 
-        [FieldsetDescription("Volatility", typeof(float))]
+        [FieldsetDescription("Volatility", typeof(double))]
         Volatility,
 
-        [FieldsetDescription("VWAP", typeof(float))]
+        [FieldsetDescription("VWAP", typeof(double))]
         VWAP,
     }
 }
